fix debian bug#493966 ia64 compilation
git-svn-id: file:///.data/svn_convert/svn@7915 f7a67266-ba00-4dfb-90a2-3eec4ba07b7d

+++ trunk/plugins/icq/icqlists.cpp
@@ -1078,7 +1078,7 @@ unsigned short ICQClient::sendRoster(unsigned short cmd, const QString &name, un
 
     snac(ICQ_SNACxFOOD_LISTS, cmd, true);
     QCString sName = name.utf8();
-    socket()->writeBuffer().pack(htons(sName.length()));
+    socket()->writeBuffer().pack(static_cast<unsigned short>( htons(sName.length()) ) );
     socket()->writeBuffer().pack(sName.data(), sName.length());
     socket()->writeBuffer()
     << grp_id
